build(dockefile fix): changed a dockerfile
BREAKING CHANGE: devel version

+++ build/package/docker-entrypoint.sh
@@ -24,7 +24,7 @@ if [ "$1" = 'start' ]; then
         python manage.py collectstatic --noinput  # Collect static files       
         echo "Installing creating Superuser (admin:chageme)..."
         python -c "import django; django.setup(); from django.contrib.auth.management.commands.createsuperuser import get_user_model; get_user_model()._default_manager.db_manager('$DJANGO_DB_NAME').create_superuser( username='$DJANGO_SU_NAME', email='$DJANGO_SU_EMAIL', password='$DJANGO_SU_PASSWORD')" # buildkit 418kB buildkit.dockerfile.v0
-        mv /Procfile .
+        [ -f /Procfile ] && mv /Procfile .
         chown -R $USER:$USER /code
     fi
     exec honcho "$@"
